requirements.txt
requests
PyYAML
Pillow
playwright
playwright-stealth
fake-useragent 